// src/constants/types.ts
export type Link = {
   title: string;
   path: string;
};

export type Company = {
   imgSrc: string;
   alt: string;
};

export type Highlights = {
   buttonText: string;
   header: string;
   imgAlign: string;
   imgSrc: string;
   paragraph: string;
   title: string;
};

export type Testimonial = {
   feedback: string;
   imgSrc: string;
   name: string;
};